requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-migrate==4.0.5
werkzeug==2.3.7
python-dotenv==1.0.0
fpdf==1.7.2
pydrive2==1.19.0
pandas==2.1.1
openpyxl==3.1.2
email-validator==2.0.0
flask-wtf==1.1.1
sqlalchemy==2.0.23
gunicorn==21.2.0
Pillow==9.5.0
python-dateutil==2.8.2